crop-disease-backend/pyproject.toml
# Crop Disease Detection Backend
version = "1.0.0"
description = "Production-ready crop disease detection API using hybrid two-stage AI pipeline"
authors = ["Senior AI Engineer"]
license = "MIT"

[dependencies]
python = "^3.8"

[tool.poetry.dependencies]
fastapi = "^0.104.1"
uvicorn = {version = "^0.24.0", extras = ["standard"]}
ultralytics = "^8.0.227"
tensorflow-cpu = "^2.20.0"
pillow = "^10.1.0"
python-multipart = "^0.0.6"
numpy = "^1.24.3"
opencv-python-headless = "^4.8.1.78"
requests = "^2.31.0"
pydantic = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"